NATIVE-44: Rename withError to withPayloadProvider
